Split Navigator and NavigatorImpl
